web_app/my_web_app/requirements.txt
azure-core==1.30.2
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.20.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Flask==3.0.3
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
msal==1.29.0
msal-extensions==1.2.0
portalocker==2.10.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
